third_party/WebKit/LayoutTests/xr/requestDevice_two_devices.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../xr/resources/xr-device-mocking.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<script src="../xr/resources/test-constants.js"></script>
<script>
let fakeDevices = fakeXRDevices();
promise_test( (t) => {
  // TODO(offenwanger) When we know how to we want to select which device to
  // return, update this test.
  setFakeDevices([fakeDevices["FakeGooglePixelPhone"], fakeDevices["FakeMagicWindowOnly"]]);
  return navigator.xr.requestDevice().then( (device) => {
    t.step(() => {
      assert_true(device != null);
      assert_true(device instanceof XRDevice);
    });

    // Repeated calls to requestDevice should yeild the same device each time.
    navigator.xr.requestDevice().then( (device2) => {
      t.step(() => {
        assert_true(device2 != null);
        assert_true(device2 instanceof XRDevice);
        assert_equals(device, device2);
      });
    });
  });
}, "navigator.xr.requestDevice properly returns one device when there are two");
</script>
